Refactor Extensions folder (#436)
