+++ SkeletonViewCore/Sources/Collections/SkeletonCollectionDelegate.swift
@@ -56,7 +56,7 @@ extension SkeletonCollectionDelegate: UICollectionViewDelegate { }
 extension SkeletonCollectionDelegate {
     private func skeletonViewIfContainerSkeletonIsActive(container: UIView, view: UIView) {
         guard container.isSkeletonActive,
-              let skeletonConfig = container.currentSkeletonConfig else {
+              let skeletonConfig = container._currentSkeletonConfig else {
             return
         }
