+++ SkeletonViewCore/Sources/Internal/FoundationExtensions/Notification+Extensions.swift
@@ -1,0 +1,22 @@
+//
+//  Copyright SkeletonView. All Rights Reserved.
+//
+//  Licensed under the MIT License (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      https://opensource.org/licenses/MIT
+//
+//  Notification+Extensions.swift
+//
+//  Created by Juanpe Catalán on 18/8/21.
+
+import UIKit
+
+extension Notification.Name {
+    
+    static let applicationDidBecomeActiveNotification = UIApplication.didBecomeActiveNotification
+    static let applicationWillTerminateNotification = UIApplication.willTerminateNotification
+    static let applicationDidEnterForegroundNotification = UIApplication.didEnterBackgroundNotification
+    
+}
