+++ SkeletonViewCore/Sources/Internal/UIKitExtensions/UIColor+Skeleton.swift
@@ -3,52 +3,6 @@
 import UIKit
 
 // codebeat:disable[TOO_MANY_IVARS]
-extension UIColor {
-    convenience init(_ hex: UInt) {
-        self.init(
-            red: CGFloat((hex & 0xFF0000) >> 16) / 255.0,
-            green: CGFloat((hex & 0x00FF00) >> 8) / 255.0,
-            blue: CGFloat(hex & 0x0000FF) / 255.0,
-            alpha: CGFloat(1.0)
-        )
-    }
-    
-    func isLight() -> Bool {
-        guard let components = cgColor.components,
-            components.count >= 3 else { return false }
-        let brightness = ((components[0] * 299) + (components[1] * 587) + (components[2] * 114)) / 1000
-        return !(brightness < 0.5)
-    }
-    
-    public var complementaryColor: UIColor {
-        if #available(iOS 13, tvOS 13, *) {
-            return UIColor { _ in
-                return self.isLight() ? self.darker : self.lighter
-            }
-        } else {
-            return isLight() ? darker : lighter
-        }
-    }
-    
-    public var lighter: UIColor {
-        adjust(by: 1.35)
-    }
-    
-    public var darker: UIColor {
-        adjust(by: 0.94)
-    }
-    
-    func adjust(by percent: CGFloat) -> UIColor {
-        var h: CGFloat = 0, s: CGFloat = 0, b: CGFloat = 0, a: CGFloat = 0
-        getHue(&h, saturation: &s, brightness: &b, alpha: &a)
-        return UIColor(hue: h, saturation: s, brightness: b * percent, alpha: a)
-    }
-    
-    func makeGradient() -> [UIColor] {
-        [self, self.complementaryColor, self]
-    }
-}
-
 public extension UIColor {
     
     static var greenSea     = UIColor(0x16a085)
@@ -87,5 +41,58 @@ public extension UIColor {
             return .clouds
         }
     }
+    
+    var complementaryColor: UIColor {
+        if #available(iOS 13, tvOS 13, *) {
+            return UIColor { [weak self] _ in
+                guard let self = self else {
+                    fatalError("Complementary color can't be calculated because original color is `nil`")
+                }
+                
+                return self.isLight ? self.darker : self.lighter
+            }
+        } else {
+            return isLight ? darker : lighter
+        }
+    }
+    
+    var lighter: UIColor {
+        adjust(by: 1.35)
+    }
+    
+    var darker: UIColor {
+        adjust(by: 0.94)
+    }
+    
+}
+
+extension UIColor {
+    
+    convenience init(_ hex: UInt) {
+        self.init(
+            red: CGFloat((hex & 0xFF0000) >> 16) / 255.0,
+            green: CGFloat((hex & 0x00FF00) >> 8) / 255.0,
+            blue: CGFloat(hex & 0x0000FF) / 255.0,
+            alpha: CGFloat(1.0)
+        )
+    }
+    
+    var isLight: Bool {
+        guard let components = cgColor.components,
+            components.count >= 3 else { return false }
+        let brightness = ((components[0] * 299) + (components[1] * 587) + (components[2] * 114)) / 1000
+        return !(brightness < 0.5)
+    }
+    
+    func adjust(by percent: CGFloat) -> UIColor {
+        var h: CGFloat = 0, s: CGFloat = 0, b: CGFloat = 0, a: CGFloat = 0
+        getHue(&h, saturation: &s, brightness: &b, alpha: &a)
+        return UIColor(hue: h, saturation: s, brightness: b * percent, alpha: a)
+    }
+    
+    func makeGradient() -> [UIColor] {
+        [self, self.complementaryColor, self]
+    }
+    
 }
 // codebeat:enable[TOO_MANY_IVARS]
