+++ SkeletonViewCore/Sources/Internal/UIKitExtensions/UIView+Extensions.swift
@@ -1,0 +1,101 @@
+//
+//  Copyright SkeletonView. All Rights Reserved.
+//
+//  Licensed under the MIT License (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      https://opensource.org/licenses/MIT
+//
+//  UIView+Extensions.swift
+//
+//  Created by Juanpe Catalán on 18/8/21.
+
+import UIKit
+
+extension UIView {
+
+    /// Flags
+    
+    var isSuperviewAStackView: Bool {
+        superview is UIStackView
+    }
+    
+    var isSkeletonActive: Bool {
+        return _status == .on || subviewsSkeletonables.contains(where: { $0.isSkeletonActive })
+    }
+    
+    var isRTL: Bool {
+        if #available(iOS 10.0, *), #available(tvOS 10.0, *) {
+            return effectiveUserInterfaceLayoutDirection == .rightToLeft
+        } else {
+            return false
+        }
+    }
+    
+    /// Math
+    
+    var definedMaxBounds: CGRect {
+        if let parentStackView = (superview as? UIStackView) {
+            var origin: CGPoint = .zero
+            switch parentStackView.alignment {
+            case .trailing:
+                origin.x = definedMaxWidth
+            default:
+                break
+            }
+            return CGRect(origin: origin, size: definedMaxSize)
+        }
+        return CGRect(origin: .zero, size: definedMaxSize)
+    }
+    
+    var definedMaxSize: CGSize {
+        CGSize(width: definedMaxWidth, height: definedMaxHeight)
+    }
+    
+    var definedMaxWidth: CGFloat {
+        let constraintsMaxWidth = widthConstraints
+            .map { $0.constant }
+            .max() ?? 0
+
+        return max(frame.size.width, constraintsMaxWidth)
+    }
+    
+    var definedMaxHeight: CGFloat {
+        let constraintsMaxHeight = heightConstraints
+            .map { $0.constant }
+            .max() ?? 0
+
+        return max(frame.size.height, constraintsMaxHeight)
+    }
+    
+    /// Autolayout
+    
+    var widthConstraints: [NSLayoutConstraint] {
+        nonContentSizeLayoutConstraints.filter { $0.firstAttribute == NSLayoutConstraint.Attribute.width }
+    }
+    
+    var heightConstraints: [NSLayoutConstraint] {
+        nonContentSizeLayoutConstraints.filter { $0.firstAttribute == NSLayoutConstraint.Attribute.height }
+    }
+    
+    var skeletonHeightConstraints: [NSLayoutConstraint] {
+        nonContentSizeLayoutConstraints.filter {
+            $0.firstAttribute == NSLayoutConstraint.Attribute.height
+                && $0.identifier?.contains("SkeletonView.Constraint.Height") ?? false
+        }
+    }
+    
+    @discardableResult
+    func setHeight(equalToConstant constant: CGFloat) -> NSLayoutConstraint {
+        let heightConstraint = heightAnchor.constraint(equalToConstant: constant)
+        heightConstraint.identifier = "SkeletonView.Constraint.Height.\(constant)"
+        NSLayoutConstraint.activate([heightConstraint])
+        return heightConstraint
+    }
+    
+    var nonContentSizeLayoutConstraints: [NSLayoutConstraint] {
+        constraints.filter({ "\(type(of: $0))" != "NSContentSizeLayoutConstraint" })
+    }
+    
+}
