+++ SkeletonViewCore/Sources/SkeletonFlow.swift
@@ -15,21 +15,21 @@ protocol SkeletonFlowDelegate: AnyObject {
 
 class SkeletonFlowHandler: SkeletonFlowDelegate {
     func willBeginShowingSkeletons(rootView: UIView) {
-        NotificationCenter.default.post(name: .willBeginShowingSkeletons, object: rootView, userInfo: nil)
-        rootView.addAppNotificationsObservers()
+        NotificationCenter.default.post(name: .skeletonWillAppearNotification, object: rootView, userInfo: nil)
+        rootView.startObservingAppLifecycleNotifications()
     }
 
     func didShowSkeletons(rootView: UIView) {
         printSkeletonHierarchy(in: rootView)
-        NotificationCenter.default.post(name: .didShowSkeletons, object: rootView, userInfo: nil)
+        NotificationCenter.default.post(name: .skeletonWillAppearNotification, object: rootView, userInfo: nil)
     }
 
     func willBeginUpdatingSkeletons(rootView: UIView) {
-        NotificationCenter.default.post(name: .willBeginUpdatingSkeletons, object: rootView, userInfo: nil)
+        NotificationCenter.default.post(name: .skeletonWillAppearNotification, object: rootView, userInfo: nil)
     }
 
     func didUpdateSkeletons(rootView: UIView) {
-        NotificationCenter.default.post(name: .didUpdateSkeletons, object: rootView, userInfo: nil)
+        NotificationCenter.default.post(name: .skeletonWillAppearNotification, object: rootView, userInfo: nil)
     }
 
     func willBeginLayingSkeletonsIfNeeded(rootView: UIView) {
@@ -39,21 +39,12 @@ class SkeletonFlowHandler: SkeletonFlowDelegate {
     }
 
     func willBeginHidingSkeletons(rootView: UIView) {
-        NotificationCenter.default.post(name: .willBeginHidingSkeletons, object: rootView, userInfo: nil)
-        rootView.removeAppNoticationsObserver()
+        NotificationCenter.default.post(name: .skeletonWillDisappearNotification, object: rootView, userInfo: nil)
+        rootView.stopObservingAppLifecycleNotications()
     }
 
     func didHideSkeletons(rootView: UIView) {
-        rootView.flowDelegate = nil
-        NotificationCenter.default.post(name: .didHideSkeletons, object: rootView, userInfo: nil)
+        rootView._flowDelegate = nil
+        NotificationCenter.default.post(name: .skeletonDidDisappearNotification, object: rootView, userInfo: nil)
     }
-}
-
-public extension Notification.Name {
-    static let willBeginShowingSkeletons = Notification.Name("willBeginShowingSkeletons")
-    static let didShowSkeletons = Notification.Name("didShowSkeletons")
-    static let willBeginUpdatingSkeletons = Notification.Name("willBeginUpdatingSkeletons")
-    static let didUpdateSkeletons = Notification.Name("didUpdateSkeletons")
-    static let willBeginHidingSkeletons = Notification.Name("willBeginHidingSkeletons")
-    static let didHideSkeletons = Notification.Name("didHideSkeletons")
 }
