+++ SkeletonViewCore/Sources/SkeletonView.swift
@@ -9,7 +9,7 @@ public extension UIView {
     ///   - color: The color of the skeleton. Defaults to `SkeletonAppearance.default.tintColor`.
     ///   - transition: The style of the transition when the skeleton appears. Defaults to `.crossDissolve(0.25)`.
     func showSkeleton(usingColor color: UIColor = SkeletonAppearance.default.tintColor, transition: SkeletonTransitionStyle = .crossDissolve(0.25)) {
-        delayedShowSkeletonWorkItem?.cancel()
+        _delayedShowSkeletonWorkItem?.cancel()
         let config = SkeletonConfig(type: .solid, colors: [color], transition: transition)
         showSkeleton(skeletonConfig: config)
     }
@@ -22,14 +22,14 @@ public extension UIView {
     ///   - delay: The amount of time (measured in seconds) to wait before show the skeleton.
     ///   - transition: The style of the transition when the skeleton appears. Defaults to `.crossDissolve(0.25)`.
     func showSkeleton(usingColor color: UIColor = SkeletonAppearance.default.tintColor, animated: Bool = true, delay: TimeInterval, transition: SkeletonTransitionStyle = .crossDissolve(0.25)) {
-        delayedShowSkeletonWorkItem?.cancel()
+        _delayedShowSkeletonWorkItem?.cancel()
         
-        delayedShowSkeletonWorkItem = DispatchWorkItem { [weak self] in
+        _delayedShowSkeletonWorkItem = DispatchWorkItem { [weak self] in
             let config = SkeletonConfig(type: .solid, colors: [color], animated: animated, transition: transition)
             self?.showSkeleton(skeletonConfig: config)
         }
         
-        DispatchQueue.main.asyncAfter(deadline: .now() + delay, execute: delayedShowSkeletonWorkItem!)
+        DispatchQueue.main.asyncAfter(deadline: .now() + delay, execute: _delayedShowSkeletonWorkItem!)
     }
     
     /// Shows the gradient skeleton without animation using the view that calls this method as root view.
@@ -38,7 +38,7 @@ public extension UIView {
     ///   - gradient: The gradient of the skeleton. Defaults to `SkeletonAppearance.default.gradient`.
     ///   - transition: The style of the transition when the skeleton appears. Defaults to `.crossDissolve(0.25)`.
     func showGradientSkeleton(usingGradient gradient: SkeletonGradient = SkeletonAppearance.default.gradient, transition: SkeletonTransitionStyle = .crossDissolve(0.25)) {
-        delayedShowSkeletonWorkItem?.cancel()
+        _delayedShowSkeletonWorkItem?.cancel()
         let config = SkeletonConfig(type: .gradient, colors: gradient.colors, transition: transition)
         showSkeleton(skeletonConfig: config)
     }
@@ -56,14 +56,14 @@ public extension UIView {
         delay: TimeInterval,
         transition: SkeletonTransitionStyle = .crossDissolve(0.25)
     ) {
-        delayedShowSkeletonWorkItem?.cancel()
+        _delayedShowSkeletonWorkItem?.cancel()
         
-        delayedShowSkeletonWorkItem = DispatchWorkItem { [weak self] in
+        _delayedShowSkeletonWorkItem = DispatchWorkItem { [weak self] in
             let config = SkeletonConfig(type: .gradient, colors: gradient.colors, animated: animated, transition: transition)
             self?.showSkeleton(skeletonConfig: config)
         }
         
-        DispatchQueue.main.asyncAfter(deadline: .now() + delay, execute: delayedShowSkeletonWorkItem!)
+        DispatchQueue.main.asyncAfter(deadline: .now() + delay, execute: _delayedShowSkeletonWorkItem!)
     }
     
     /// Shows the animated skeleton using the view that calls this method as root view.
@@ -75,7 +75,7 @@ public extension UIView {
     ///   - animation: The animation of the skeleton. Defaults to `nil`.
     ///   - transition: The style of the transition when the skeleton appears. Defaults to `.crossDissolve(0.25)`.
     func showAnimatedSkeleton(usingColor color: UIColor = SkeletonAppearance.default.tintColor, animation: SkeletonLayerAnimation? = nil, transition: SkeletonTransitionStyle = .crossDissolve(0.25)) {
-        delayedShowSkeletonWorkItem?.cancel()
+        _delayedShowSkeletonWorkItem?.cancel()
         let config = SkeletonConfig(type: .solid, colors: [color], animated: true, animation: animation, transition: transition)
         showSkeleton(skeletonConfig: config)
     }
@@ -89,7 +89,7 @@ public extension UIView {
     ///   - animation: The animation of the skeleton. Defaults to `nil`.
     ///   - transition: The style of the transition when the skeleton appears. Defaults to `.crossDissolve(0.25)`.
     func showAnimatedGradientSkeleton(usingGradient gradient: SkeletonGradient = SkeletonAppearance.default.gradient, animation: SkeletonLayerAnimation? = nil, transition: SkeletonTransitionStyle = .crossDissolve(0.25)) {
-        delayedShowSkeletonWorkItem?.cancel()
+        _delayedShowSkeletonWorkItem?.cancel()
         let config = SkeletonConfig(type: .gradient, colors: gradient.colors, animated: true, animation: animation, transition: transition)
         showSkeleton(skeletonConfig: config)
     }
@@ -115,13 +115,13 @@ public extension UIView {
     }
 
     func layoutSkeletonIfNeeded() {
-        flowDelegate?.willBeginLayingSkeletonsIfNeeded(rootView: self)
+        _flowDelegate?.willBeginLayingSkeletonsIfNeeded(rootView: self)
         recursiveLayoutSkeletonIfNeeded(root: self)
     }
     
     func hideSkeleton(reloadDataAfter reload: Bool = true, transition: SkeletonTransitionStyle = .crossDissolve(0.25)) {
-        delayedShowSkeletonWorkItem?.cancel()
-        flowDelegate?.willBeginHidingSkeletons(rootView: self)
+        _delayedShowSkeletonWorkItem?.cancel()
+        _flowDelegate?.willBeginHidingSkeletons(rootView: self)
         recursiveHideSkeleton(reloadDataAfter: reload, transition: transition, root: self)
     }
     
@@ -148,14 +148,14 @@ extension UIView {
 
     @objc func skeletonTraitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
         skeletonTraitCollectionDidChange(previousTraitCollection)
-        guard isSkeletonable, isSkeletonActive, let config = currentSkeletonConfig else { return }
+        guard isSkeletonable, isSkeletonActive, let config = _currentSkeletonConfig else { return }
         updateSkeleton(skeletonConfig: config)
     }
     
     func showSkeleton(skeletonConfig config: SkeletonConfig) {
-        isSkeletonAnimated = config.animated
-        flowDelegate = SkeletonFlowHandler()
-        flowDelegate?.willBeginShowingSkeletons(rootView: self)
+        _isSkeletonAnimated = config.animated
+        _flowDelegate = SkeletonFlowHandler()
+        _flowDelegate?.willBeginShowingSkeletons(rootView: self)
         recursiveShowSkeleton(skeletonConfig: config, root: self)
     }
 
@@ -164,7 +164,7 @@ extension UIView {
             isHidden = true
         }
         guard isSkeletonable && !isSkeletonActive else { return }
-        currentSkeletonConfig = config
+        _currentSkeletonConfig = config
         swizzleLayoutSubviews()
         swizzleTraitCollectionDidChange()
         addDummyDataSourceIfNeeded()
@@ -175,7 +175,7 @@ extension UIView {
         }
 
         if let root = root {
-            flowDelegate?.didShowSkeletons(rootView: root)
+            _flowDelegate?.didShowSkeletons(rootView: root)
         }
     }
 
@@ -188,17 +188,17 @@ extension UIView {
     }
 
     func updateSkeleton(skeletonConfig config: SkeletonConfig) {
-        isSkeletonAnimated = config.animated
-        flowDelegate?.willBeginUpdatingSkeletons(rootView: self)
+        _isSkeletonAnimated = config.animated
+        _flowDelegate?.willBeginUpdatingSkeletons(rootView: self)
         recursiveUpdateSkeleton(skeletonConfig: config, root: self)
     }
 
     private func recursiveUpdateSkeleton(skeletonConfig config: SkeletonConfig, root: UIView? = nil) {
         guard isSkeletonActive else { return }
-        currentSkeletonConfig = config
+        _currentSkeletonConfig = config
         updateDummyDataSourceIfNeeded()
         subviewsSkeletonables.recursiveSearch(leafBlock: {
-            if let skeletonLayer = skeletonLayer,
+            if let skeletonLayer = _skeletonLayer,
                 skeletonLayer.type != config.type {
                 removeSkeletonLayer()
                 addSkeletonLayer(skeletonConfig: config)
@@ -210,7 +210,7 @@ extension UIView {
         }
 
         if let root = root {
-            flowDelegate?.didUpdateSkeletons(rootView: root)
+            _flowDelegate?.didUpdateSkeletons(rootView: root)
         }
     }
 
@@ -218,7 +218,7 @@ extension UIView {
         subviewsSkeletonables.recursiveSearch(leafBlock: {
             guard isSkeletonable, isSkeletonActive else { return }
             layoutSkeletonLayerIfNeeded()
-            if let config = currentSkeletonConfig, config.animated, !isSkeletonAnimated {
+            if let config = _currentSkeletonConfig, config.animated, !_isSkeletonAnimated {
                 startSkeletonAnimation(config.animation)
             }
         }) { subview in
@@ -226,7 +226,7 @@ extension UIView {
         }
 
         if let root = root {
-            flowDelegate?.didLayoutSkeletonsIfNeeded(rootView: root)
+            _flowDelegate?.didLayoutSkeletonsIfNeeded(rootView: root)
         }
     }
 
@@ -235,7 +235,7 @@ extension UIView {
         if isHiddenWhenSkeletonIsActive {
             isHidden = false
         }
-        currentSkeletonConfig?.transition = transition
+        _currentSkeletonConfig?.transition = transition
         unSwizzleLayoutSubviews()
         unSwizzleTraitCollectionDidChange()
         removeDummyDataSourceIfNeeded(reloadAfter: reload)
@@ -247,24 +247,24 @@ extension UIView {
         }
         
         if let root = root {
-            flowDelegate?.didHideSkeletons(rootView: root)
+            _flowDelegate?.didHideSkeletons(rootView: root)
         }
     }
     
     private func startSkeletonLayerAnimationBlock(_ anim: SkeletonLayerAnimation? = nil) -> VoidBlock {
         return {
-            self.isSkeletonAnimated = true
-            guard let layer = self.skeletonLayer else { return }
+            self._isSkeletonAnimated = true
+            guard let layer = self._skeletonLayer else { return }
             layer.start(anim) { [weak self] in
-                self?.isSkeletonAnimated = false
+                self?._isSkeletonAnimated = false
             }
         }
     }
     
     private var stopSkeletonLayerAnimationBlock: VoidBlock {
         return {
-            self.isSkeletonAnimated = false
-            guard let layer = self.skeletonLayer else { return }
+            self._isSkeletonAnimated = false
+            guard let layer = self._skeletonLayer else { return }
             layer.stopAnimation()
         }
     }
@@ -324,7 +324,7 @@ extension UIView {
             .build()
             else { return }
         
-        self.skeletonLayer = skeletonLayer
+        self._skeletonLayer = skeletonLayer
         layer.insertSkeletonLayer(
             skeletonLayer,
             atIndex: UInt32.max,
@@ -340,11 +340,11 @@ extension UIView {
                 self.startSkeletonAnimation(config.animation)
             }
         }
-        status = .on
+        _status = .on
     }
     
     func updateSkeletonLayer(skeletonConfig config: SkeletonConfig) {
-        guard let skeletonLayer = skeletonLayer else { return }
+        guard let skeletonLayer = _skeletonLayer else { return }
         skeletonLayer.update(usingColors: config.colors)
         if config.animated {
             startSkeletonAnimation(config.animation)
@@ -354,19 +354,19 @@ extension UIView {
     }
 
     func layoutSkeletonLayerIfNeeded() {
-        guard let skeletonLayer = skeletonLayer else { return }
+        guard let skeletonLayer = _skeletonLayer else { return }
         skeletonLayer.layoutIfNeeded()
     }
     
     func removeSkeletonLayer() {
         guard isSkeletonActive,
-            let skeletonLayer = skeletonLayer,
-            let transitionStyle = currentSkeletonConfig?.transition else { return }
+            let skeletonLayer = _skeletonLayer,
+            let transitionStyle = _currentSkeletonConfig?.transition else { return }
         skeletonLayer.stopAnimation()
-        status = .off
+        _status = .off
         skeletonLayer.removeLayer(transition: transitionStyle) {
-            self.skeletonLayer = nil
-            self.currentSkeletonConfig = nil
+            self._skeletonLayer = nil
+            self._currentSkeletonConfig = nil
         }
     }
 }
